+++ src/posix/termios.rs
@@ -8,7 +8,7 @@ use std::os::unix::prelude::*;
 
 cfg_if! {
     if #[cfg(any(
-        target_os = "dragonflybsd",
+        target_os = "dragonfly",
         target_os = "freebsd",
         target_os = "ios",
         target_os = "macos",
@@ -59,7 +59,7 @@ pub(crate) fn get_termios(fd: RawFd) -> Result<Termios> {
 }
 
 #[cfg(any(
-    target_os = "dragonflybsd",
+    target_os = "dragonfly",
     target_os = "freebsd",
     target_os = "netbsd",
     target_os = "openbsd",
@@ -101,7 +101,7 @@ pub(crate) fn set_termios(fd: RawFd, termios: &libc::termios, baud_rate: u32) ->
     let res = unsafe { libc::tcsetattr(fd, libc::TCSANOW, termios) };
     nix::errno::Errno::result(res)?;
 
-    // Note: attempting to set the baud rate on a pseudo terminal via this ioctl call will faill
+    // Note: attempting to set the baud rate on a pseudo terminal via this ioctl call will fail
     // with the `ENOTTY` error.
     if baud_rate > 0 {
         crate::posix::ioctl::iossiospeed(fd, &(baud_rate as libc::speed_t))?;
@@ -111,7 +111,7 @@ pub(crate) fn set_termios(fd: RawFd, termios: &libc::termios, baud_rate: u32) ->
 }
 
 #[cfg(any(
-    target_os = "dragonflybsd",
+    target_os = "dragonfly",
     target_os = "freebsd",
     target_os = "netbsd",
     target_os = "openbsd",
@@ -222,7 +222,7 @@ pub(crate) fn set_baud_rate(termios: &mut Termios, baud_rate: u32) {
 
 // BSDs use the baud rate as the constant value so there's no translation necessary
 #[cfg(any(
-    target_os = "dragonflybsd",
+    target_os = "dragonfly",
     target_os = "freebsd",
     target_os = "netbsd",
     target_os = "openbsd"